add LocalStorageKey; rework LocalStorage API (#30)
# add LocalStorageKey; rework LocalStorage API

## :recycle: Current situation & Problem
`LocalStorage` currently identifies entries in the storage via either
optional, raw untyped strings, or via the static type of the
to-be-stored/loaded value.
While this approach does lead to a nice-looking API, it is problematic
for several reasons:
- identifying stored entries using string literals makes it easy to
misspel something, or to accidentally use the wrong key
- the fact that there's no static coupling between a key and the type
it's used to store/load, means that there's nothing that would prevent
someone from acidentally decoding a value into an incorrect type (i.e.,
a different type than what was encoded into the `LocalStorage` entry)
- the fact that parameters such as e.g. the encoder/decoder or the
storage setting, both of which determine how a value is stored/loaded,
need to be explicitly specified at both calls to `store()` as well as
calls to `read()`, means that it is easy to mismatch these parameters
- eg: if you use a `PropertyListEncoder` when storing some value into
the `LocalStorage`, you need to make sure that you also use the matching
`PropertyListDecoder` when reading the stored value.
- since the current API defaults both the encoder in `store()` and the
decoder in `read()` to the `JSONEncoder` (and `JSONDecoder,
respectively), makes it too easy to miss the parameter somewhere
- the same applies to the `storageSetting` parameter: all matching calls
to `store()` and `read()` that operate on the same storage key must
specify the same storage setting.
- this is currently not enforced, and with the current API there is no
way of enforcing this
- the fact that storage keys are optional (if the key is nil (the
default), the unqualified typename of the to-be-stored/loaded type will
be used instead) is also likely to lead to unexpected issues:
- since the unqualified typename doesn't take nesting / access control
into account, it can easily happen that two calls to `store()`/`read()`
that operate on completely different types (potentially even in
completely different modules!) end up operating on the same LocalStorage
entry, because the unqualified typename will be the same in both cases

This PR attempts to address these issues, by restructuring and
redesigning the `LocalStorage` API:
- Entries in the storage are no longer identified by raw strings which
are passed to the `store()`/`load()` calls
- Entries are no longer identified by the typename of the value being
stored/loaded
- We instead have a `LocalStorageKey<Value>` type, which are intended to
be long-lived objects that are used to access the `LocalStorage`
- the `LocalStorageKey` encapsulates all information required to store
data into the storage / load data from it
- since every store/load for an entry will use the same
`LocalStorageKey` instance, we can ensure that every store/load for the
same entry will use matching decoders/encoders, and will use the same
storageSetting.
- it also improves the typing, by associating every storage key with its
underlying stored type
- in order to prevent race conditions and other potential concurrency
issues, the `LocalStorageKey` also acts as a read-write lock when
accessing the storage entry identified by it
- In addition to these changes, the `LocalStorageKey`-based API also
makes it easier to support storing additional currently not supported
types, such as e.g. `NSSecureCoding`-compliant types
- **Note** this PR removes support for `CodableWithConfiguration`, which
previously was supported but (AFAICT) not used.
- **Note**: `SpeziSecureStorage` is currently untouched; it might make
sense to have something similar there as well, but it'd of course
require a separate, dedicated key type and would be out-of-scope for
this PR.
- This PR also adds a small convenience property wrapper,
`@LocalStorageEntry`, which makes it easier to use the `LocalStorage`
API from within a SwiftUI View, and is e.g. able to observe a key's
value and update the view in response.


## :gear: Release Notes 
- Completely reworked the `LocalStorage` API:
- Values are stored/loaded via `LocalStorageKey`s, which define an entry
in the storage.
- `LocalStorageKey`s need to explicitly specify their underlying raw
keys (a String), and are encouraged to use reverse-DNS-notation in order
to avoid collisions


## :books: Documentation
The documentation has been adapted to the new API.

## :white_check_mark: Testing
The tests have been adapted to the new API.


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).